+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CraftHelper updates
 
-## 1.5.6 ()
+## 1.5.6 (18.04.2022)
 
 ### New features